Reorganize integration tests for better clarity and trackability of CLI tests
