+++ lib/http-bridge/cardano-wallet-http-bridge.cabal
@@ -172,10 +172,13 @@ test-suite integration
       Test.Integration.Framework.DSL
       Test.Integration.Framework.Request
       Test.Integration.Framework.TestData
-      Test.Integration.Scenario.Addresses
-      Test.Integration.Scenario.Transactions
-      Test.Integration.Scenario.Wallets
-      Test.Integration.Scenario.CLISpec
+      Test.Integration.Scenario.API.Addresses
+      Test.Integration.Scenario.API.Transactions
+      Test.Integration.Scenario.API.Wallets
+      Test.Integration.Scenario.CLI.Addresses
+      Test.Integration.Scenario.CLI.Mnemonics
+      Test.Integration.Scenario.CLI.Transactions
+      Test.Integration.Scenario.CLI.Wallets
 
 benchmark restore
   default-language:
