+++ lib/http-bridge/test/integration/Test/Integration/Scenario/CLI/Wallets.hs
@@ -2,31 +2,22 @@
 {-# LANGUAGE OverloadedLabels #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Test.Integration.Scenario.CLISpec
-    ( specNoCluster
-    , specWithCluster
+module Test.Integration.Scenario.CLI.Wallets
+    ( spec
     ) where
 
 import Prelude
 
 import Cardano.Wallet.Api.Types
-    ( ApiAddress, ApiTransaction, ApiWallet, getApiT )
-import Cardano.Wallet.HttpBridge.Compatibility
-    ( HttpBridge )
-import Cardano.Wallet.Primitive.Types
-    ( encodeAddress )
+    ( ApiWallet )
 import Control.Monad
     ( forM_ )
 import Data.Functor
     ( ($>) )
 import Data.Generics.Internal.VL.Lens
-    ( view, (^.) )
-import Data.List
-    ( length )
+    ( view )
 import Data.Proxy
     ( Proxy (..) )
-import Data.Text
-    ( Text )
 import System.Command
     ( Exit (..), Stderr (..), Stdout (..) )
 import System.Exit
@@ -37,21 +28,13 @@ import Test.Hspec.Expectations.Lifted
     ( shouldBe, shouldContain, shouldNotContain )
 import Test.Integration.Framework.DSL
     ( Context (..)
-    , Payload (..)
-    , cardanoWalletCLI
-    , cardanoWalletLauncherCLI
     , createWalletViaCLI
     , deleteWalletViaCLI
     , emptyWallet
     , expectValidJSON
-    , fixtureWallet
     , generateMnemonicsViaCLI
-    , getAddresses
     , getWalletViaCLI
-    , listAddressesViaCLI
     , listWalletsViaCLI
-    , postTransactionViaCLI
-    , unsafeRequest
     , updateWalletViaCLI
     , walletId
     )
@@ -60,67 +43,8 @@ import Test.Integration.Framework.TestData
 
 import qualified Data.Text as T
 
-version :: Text
-version = "2019.5.24"
-
-specNoCluster :: SpecWith ()
-specNoCluster = do
-    it "CLI - Shows version" $  do
-        (Exit c, Stdout out) <- cardanoWalletCLI ["--version"]
-        let v = T.dropWhileEnd (== '\n') (T.pack out)
-        v `shouldBe` version
-        c `shouldBe` ExitSuccess
-
-    it "CLI - cardano-wallet-launcher shows help on bad argument" $  do
-        (Exit c, Stdout out) <- cardanoWalletLauncherCLI ["--bad arg"]
-        out `shouldContain` "cardano-wallet-launcher"
-        c `shouldBe` ExitFailure 1
-
-    describe "CLI - cardano-wallet-launcher shows help with" $  do
-        let test option = it option $ do
-                (Exit c, Stdout out) <- cardanoWalletLauncherCLI [option]
-                out `shouldContain` "cardano-wallet-launcher"
-                c `shouldBe` ExitSuccess
-        forM_ ["-h", "--help"] test
-
-    it "CLI - cardano-wallet shows help on bad argument" $  do
-        (Exit c, Stdout out) <- cardanoWalletCLI ["--bad arg"]
-        out `shouldContain` "Cardano Wallet CLI"
-        c `shouldBe` ExitFailure 1
-
-    describe "CLI - cardano-wallet shows help with" $  do
-        let test option = it option $ do
-                (Exit c, Stdout out) <- cardanoWalletCLI [option]
-                out `shouldContain` "Cardano Wallet CLI"
-                c `shouldBe` ExitSuccess
-        forM_ ["-h", "--help"] test
-
-    it "CLI - Can generate mnemonics with default size" $  do
-        (Exit c, Stdout out) <- generateMnemonicsViaCLI []
-        length (words out) `shouldBe` 15
-        c `shouldBe` ExitSuccess
-
-    describe "CLI - Can generate mnemonics with different sizes" $ do
-        let test size = it ("--size=" <> show size) $ do
-                (Exit c, Stdout out) <-
-                    generateMnemonicsViaCLI ["--size", show size]
-                length (words out) `shouldBe` size
-                c `shouldBe` ExitSuccess
-        forM_ [9, 12, 15, 18, 21, 24] test
-
-    describe "CLI - It can't generate mnemonics with an invalid size" $ do
-        let sizes =
-                ["15.5", "3", "6", "14", "abc", "👌", "0", "~!@#%" , "-1000", "1000"]
-        forM_ sizes $ \(size) -> it ("--size=" <> size) $ do
-            (Exit c, Stdout out, Stderr err) <-
-                generateMnemonicsViaCLI ["--size", size]
-            c `shouldBe` ExitFailure 1
-            err `shouldBe`
-                "Invalid mnemonic size. Expected one of: 9,12,15,18,21,24\n"
-            out `shouldBe` mempty
-
-specWithCluster :: SpecWith Context
-specWithCluster = do
+spec :: SpecWith Context
+spec = do
     describe "CLI1 - Can create wallet with different mnemonic sizes" $ do
         forM_ ["15", "18", "21", "24"] $ \(size) -> it size $ \_ -> do
             let walletName = "Wallet created via CLI"
@@ -220,27 +144,6 @@ specWithCluster = do
         out `shouldNotContain` "CLI Wallet"
         c `shouldBe` ExitSuccess
 
-    it "CLI - Can list addresses" $ \ctx -> do
-        walId <- emptyWallet' ctx
-        (Exit c, Stdout out, Stderr err) <- listAddressesViaCLI walId
-        err `shouldBe` "Ok.\n"
-        expectValidJSON (Proxy @[ApiAddress HttpBridge]) out
-        c `shouldBe` ExitSuccess
-
-    it "CLI - Can create transaction" $ \ctx -> do
-        wSrc <- fixtureWallet ctx
-        wDest <- emptyWallet ctx
-        (_, addr:_) <- unsafeRequest @[ApiAddress HttpBridge] ctx (getAddresses wDest) Empty
-        let addrStr =
-                encodeAddress (Proxy @HttpBridge) (getApiT $ fst $ addr ^. #id)
-        let args = T.unpack <$>
-                [ wSrc ^. walletId
-                , "--payment", "14@" <> addrStr
-                ]
-        (c, out, err) <- postTransactionViaCLI "cardano-wallet" args
-        err `shouldBe` "Please enter a passphrase: **************\nOk.\n"
-        expectValidJSON (Proxy @(ApiTransaction HttpBridge)) out
-        c `shouldBe` ExitSuccess
   where
     emptyWallet' :: Context -> IO String
     emptyWallet' = fmap (T.unpack . view walletId) . emptyWallet
